+++ dev/examples/mathbox/geom.clj
@@ -7,6 +7,7 @@
   (:require [emmy.clerk :as ec]
             [emmy.leva]
             [emmy.env :as e :refer :all]
+            [emmy.matrix]
             [emmy.viewer :as ev]
             [emmy.mathbox.plot :as p]))
 
@@ -15,9 +16,28 @@
 ^{:nextjournal.clerk/visibility {:code :hide :result :hide}}
 (ec/install!)
 
+(defn basis [m]
+  (let [[x y z] (emmy.matrix/transpose m)]
+    (-> [:<>
+         (p/vector {:tip x :arrow-size 3 :color "red"})
+         (p/vector {:tip y :arrow-size 3 :color "green"})
+         (p/vector {:tip z :arrow-size 3 :color "blue"})]
+        (ev/fragment p/scene))))
+
+(ev/with-let [!size {:y 1}]
+  (p/scene
+   (emmy.leva/controls
+    {:atom !size :schema {:y {:min 0 :max 10 :step 0.01}}})
+   (basis
+    (matrix-by-rows
+     [1 0 0]
+     [0 (ev/get !size :y) 0]
+     [1 0 2]))))
+
 (ev/with-let [!size {:size 20 :pos 1}]
-  [:<> (emmy.leva/controls
-        {:atom !size :schema {:size {:min 20 :max 100 :step 1}}})
+  [:<>
+   (emmy.leva/controls
+    {:atom !size :schema {:size {:min 20 :max 100 :step 1}}})
    (p/scene
     (p/point
      {:coords [0.5 -0.5 (ev/get !size :pos)]